Refactor markdown integration seams
