+++ packages/next/src/lib/constants.ts
@@ -63,6 +63,10 @@ export const DOT_NEXT_ALIAS = 'private-dot-next'
 export const ROOT_DIR_ALIAS = 'private-next-root-dir'
 export const APP_DIR_ALIAS = 'private-next-app-dir'
 export const MARKDOWN_COMPONENTS_ALIAS = 'next-markdown-import-source-file'
+export const MARKDOWN_INTERNAL_COMPONENT_BEHAVIOR_PROP =
+  '__nextMarkdownBehavior'
+export const MARKDOWN_INTERNAL_COMPONENT_BEHAVIOR_OMIT = 'omit'
+export const MARKDOWN_INTERNAL_COMPONENT_BEHAVIOR_PASSTHROUGH = 'passthrough'
 export const RSC_MOD_REF_PROXY_ALIAS = 'private-next-rsc-mod-ref-proxy'
 export const RSC_ACTION_VALIDATE_ALIAS = 'private-next-rsc-action-validate'
 export const RSC_ACTION_PROXY_ALIAS = 'private-next-rsc-server-reference'
