+++ test/production/adapter-config/pages/automatic-static-pages.js
@@ -1,3 +1,5 @@
+export const markdown = true
+
 export default function Pages() {
   return <p>automatic-static-pages</p>
 }
